Some reorganisation
