+++ nouns-protocol/src/lib.rs
@@ -8,6 +8,8 @@ pub use babyjubjub_ark::{Fr as BBJJ_Fr, Point as BBJJ_Ec, PrivateKey, B8 as BBJJ
 pub use noir::MAX_DEPTH;
 pub use noir::MAX_NODE_LEN;
 pub use tallier::{Tallier, TruncatedBallot};
+pub use utils::StateProof;
+pub use utils::BlockHeader;
 pub use utils::wrapper::Wrapper;
 pub use utils::VoteChoice;
 pub use voter::Voter;
