+++ nouns-protocol/src/noir/toml.rs
@@ -3,8 +3,8 @@ use babyjubjub_ark::Signature;
 use ethers::types::{Address, H256, StorageProof};
 use toml::Value;
 
-use crate::noir::{TallyProverInput, VoteProverInput, MAX_DEPTH, MAX_NODE_LEN};
-use crate::{utils::VoteChoice, BBJJ_Ec, BBJJ_Fr, BN254_Fr};
+use crate::noir::{TallyProverInput, VoteProverInput, MAX_ACCOUNT_STATE_SIZE, MAX_BLOCK_HEADER_SIZE, MAX_DEPTH, MAX_NODE_LEN};
+use crate::{utils::VoteChoice, BBJJ_Ec, BBJJ_Fr, BlockHeader, BN254_Fr, StateProof};
 
 pub trait TomlSerializable {
     fn toml(self) -> Value;
@@ -223,5 +223,65 @@ impl TomlSerializable for H256 {
 impl TomlSerializable for Address {
     fn toml(self) -> Value {
         BN254_Fr::from_be_bytes_mod_order(&<[u8; 20]>::from(self)).toml()
+    }
+}
+
+impl TomlSerializable for StateProof {
+    fn toml(self) -> toml::Value {
+        let mut map = toml::map::Map::new();
+        let depth = self.proof.len();
+
+        let path = self
+            .proof
+            .into_iter()
+            .map(|b| b.to_vec())
+            .collect::<Vec<_>>();
+
+        // Proof path needs to be an appropriately padded flat array.
+        let padded_path = path
+            .into_iter()
+            .chain({
+                let depth_excess = MAX_DEPTH - depth;
+                // Append with empty nodes to fill up to depth MAX_DEPTH
+                vec![vec![]; depth_excess]
+            })
+            .map(|mut v| {
+                let node_len = v.len();
+                let len_excess = MAX_NODE_LEN - node_len;
+                // Then pad each node up to length MAX_NODE_LEN
+                v.append(&mut vec![0; len_excess]);
+                v
+            })
+            .flatten()
+            .collect::<Vec<u8>>(); // And flatten.
+        map.insert("proof".to_string(), padded_path.toml());
+
+        let key: [u8; 20] = self.key.into();
+
+        let value_len = self.value.len();
+
+        assert!(value_len <= MAX_ACCOUNT_STATE_SIZE);
+
+        let mut value = vec![0u8; MAX_ACCOUNT_STATE_SIZE - value_len];
+        
+        value.append(&mut self.value.clone());
+
+        map.insert("key".to_string(), key.to_vec().toml());
+        map.insert("value".to_string(), value.toml());
+        map.insert("depth".to_string(), depth.toml());
+
+        toml::Value::Table(map)
+
+    }
+}
+
+impl TomlSerializable for BlockHeader {
+    fn toml(self) -> toml::Value {
+        let mut value = self.0.clone();
+        let value_len = value.len();
+        assert!(value_len <= MAX_BLOCK_HEADER_SIZE);
+        value.append(&mut vec![0; MAX_BLOCK_HEADER_SIZE - value_len]);
+
+        value.toml()
     }
 }
